requirements.txt
allure-pytest==2.6.2
allure-python-commons==2.6.2
Appium-Python-Client==0.40
atomicwrites==1.3.0
attrs==19.1.0
baidu-aip==2.2.13.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
enum34==1.1.6
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
kiwisolver==1.0.1
lxml==4.3.2
matplotlib==3.0.3
more-itertools==6.0.0
namedlist==1.7
numpy==1.16.2
opencv-python==4.0.0.21
openpyxl==2.6.1
Pillow==5.4.1
pluggy==0.9.0
py==1.8.0
PyMySQL==0.9.3
pyparsing==2.3.1
pytesseract==0.2.6
pytest==4.4.1
pytest-rerunfailures==7.0
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
